+++ tests/ControllerTest.php
@@ -149,17 +149,46 @@ class ControllerTest extends TestCase
 
     /**
      * Test an action returning a ActionResult.
+     *
+     * @dataProvider actionReturningActionResultDataProvider
+     *
+     * @param string $path               The path.
+     * @param int    $expectedStatusCode The expected status code.
+     * @param array  $expectedHeaders    The expected headers.
+     * @param string $expectedContent    The expected content.
      */
-    public function testActionReturningActionResult()
+    public function testActionReturningActionResult(string $path, int $expectedStatusCode, array $expectedHeaders, string $expectedContent)
     {
         $application = new BasicTestApplication(FilePath::parse('/var/www/'));
-        $request = new BasicTestRequest(Url::parse('http://www.domain.com/notfound'), new Method('GET'));
+        $request = new BasicTestRequest(Url::parse('http://www.domain.com/' . $path), new Method('GET'));
         $response = new BasicTestResponse();
         $controller = new ActionResultTestController();
-        $controller->processRequest($application, $request, $response, 'notfound');
+        $controller->processRequest($application, $request, $response, $path);
 
-        self::assertSame('Page was not found', $response->getContent());
-        self::assertSame(StatusCode::NOT_FOUND, $response->getStatusCode()->getCode());
+        self::assertSame($expectedStatusCode, $response->getStatusCode()->getCode());
+        self::assertSame($expectedHeaders, iterator_to_array($response->getHeaders()));
+        self::assertSame($expectedContent, $response->getContent());
+    }
+
+    /**
+     * Data provider for action result tests.
+     *
+     * @return array The data.
+     */
+    public function actionReturningActionResultDataProvider()
+    {
+        return [
+            ['notfound', StatusCode::NOT_FOUND, [], 'Page was not found'],
+            ['redirect', StatusCode::FOUND, ['Location' => 'http://www.domain.com/foo/bar'], ''],
+            ['permanentRedirect', StatusCode::MOVED_PERMANENTLY, ['Location' => 'https://domain.com/'], ''],
+            ['forbidden', StatusCode::FORBIDDEN, [], 'Page is forbidden'],
+            ['nocontent', StatusCode::NO_CONTENT, [], ''],
+            ['notmodified', StatusCode::NOT_MODIFIED, [], ''],
+            ['methodnotallowed', StatusCode::METHOD_NOT_ALLOWED, [], ''],
+            ['json', StatusCode::OK, ['Content-Type' => 'application/json'], '{"Foo":1,"Bar":{"Baz":2}}'],
+            ['created', StatusCode::CREATED, ['Location' => 'http://www.domain.com/created'], ''],
+            ['custom', StatusCode::MULTI_STATUS, [], 'Custom action result'],
+        ];
     }
 
     /**