+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+- Added ActionResult.
+- Deprecated AbstractActionResult.
 
 ## [2.0.0] 2018-05-14
 - **Backward compatibility break**: Replaced getException/setException methods with getThrowable/setThrowable in ErrorController.